server/python/requirements.txt
rembg[cpu]
pillow
pillow-heif
